refactor pupil add and upload to accommodate user id
