+++ admin/services/pupil-upload.service.js
@@ -3,7 +3,7 @@ const fs = require('fs-extra')
 const R = require('ramda')
 
 const pupilDataService = require('./data-access/pupil.data.service')
-const validateCSVService = require('./validate-csv.service')
+const validateCSVService = require('./pupil-validate-csv.service')
 const generateErrorCSVService = require('./generate-error-csv.service')
 const redisCacheService = require('../services/data-access/redis-cache.service')
 const redisKeyService = require('../services/redis-key.service')
@@ -12,11 +12,12 @@ const service = {}
 
 /**
  *
- * @param csvDataArray
+ * @param {Array} csvDataArray
  * @param {object} school
+ * @param {number} userId
  * @return {Promise<*>}
  */
-const onCSVReadComplete = async (csvDataArray, school) => {
+const onCSVReadComplete = async (csvDataArray, school, userId) => {
   const { pupils, csvData, headers, validationErrors, hasValidationError } = await validateCSVService.process(csvDataArray, school)
   if (hasValidationError) return ({ fileErrors: validationErrors, hasValidationError })
   // Generate csv with errors
@@ -30,7 +31,7 @@ const onCSVReadComplete = async (csvDataArray, school) => {
     })
   } else {
     // Save pupils if validation is successful
-    const result = await pupilDataService.sqlInsertMany(pupils)
+    const result = await pupilDataService.sqlInsertMany(pupils, userId)
     if (!R.prop('insertId', result)) return (new Error('No pupils were saved'))
     const pupilIds = Array.isArray(result.insertId) ? result.insertId : [result.insertId]
     return { pupilIds }
@@ -40,10 +41,12 @@ const onCSVReadComplete = async (csvDataArray, school) => {
 /**
  *
  * @param {object} school
- * @param uploadFile
+ * @param {object} uploadFile
+ * @param {number} userId
  * @return {Promise<any>}
  */
-service.upload = async (school, uploadFile) => {
+service.upload = async (school, uploadFile, userId) => {
+  if (userId === undefined) throw new Error('userId is required')
   let stream
   const pr = new Promise((resolve, reject) => {
     const csvDataArray = []
@@ -54,7 +57,7 @@ service.upload = async (school, uploadFile) => {
       })
       .on('end', async () => {
         try {
-          const response = await onCSVReadComplete(csvDataArray, school)
+          const response = await onCSVReadComplete(csvDataArray, school, userId)
           const pupilRegisterRedisKey = redisKeyService.getPupilRegisterViewDataKey(school.id)
           await redisCacheService.drop(pupilRegisterRedisKey)
           return resolve(response)
